+++ backend/ddang/src/main/java/com/ddang/ddang/auction/presentation/dto/response/SellerResponse.java
@@ -8,7 +8,7 @@ public record SellerResponse(
         Long id,
         String image,
         String nickname,
-        double reliability
+        Float reliability
 ) {
 
     public static SellerResponse from(final ReadAuctionDto auctionDto) {
