+++ backend/ddang/src/main/java/com/ddang/ddang/chat/application/dto/ReadUserInChatRoomDto.java
@@ -3,7 +3,7 @@ package com.ddang.ddang.chat.application.dto;
 import com.ddang.ddang.image.application.util.ImageIdProcessor;
 import com.ddang.ddang.user.domain.User;
 
-public record ReadUserInChatRoomDto(Long id, String name, Long profileImageId, double reliability, boolean isDeleted) {
+public record ReadUserInChatRoomDto(Long id, String name, Long profileImageId, Float reliability, boolean isDeleted) {
 
     public static ReadUserInChatRoomDto from(final User user) {
         return new ReadUserInChatRoomDto(
