+++ backend/ddang/src/main/java/com/ddang/ddang/review/application/dto/ReadUserInReviewDto.java
@@ -2,7 +2,7 @@ package com.ddang.ddang.review.application.dto;
 
 import com.ddang.ddang.user.domain.User;
 
-public record ReadUserInReviewDto(Long id, String name, Long profileImageId, double reliability, String oauthId) {
+public record ReadUserInReviewDto(Long id, String name, Long profileImageId, Float reliability, String oauthId) {
 
     public static ReadUserInReviewDto from(final User user) {
         return new ReadUserInReviewDto(
