+++ backend/ddang/src/main/java/com/ddang/ddang/user/application/dto/ReadUserDto.java
@@ -7,7 +7,7 @@ public record ReadUserDto(
         Long id,
         String name,
         Long profileImageId,
-        double reliability,
+        Float reliability,
         String oauthId,
         boolean isDeleted
 ) {
