+++ backend/ddang/src/main/java/com/ddang/ddang/user/presentation/dto/response/ReadUserResponse.java
@@ -5,7 +5,7 @@ import com.ddang.ddang.image.presentation.util.ImageUrlCalculator;
 import com.ddang.ddang.user.application.dto.ReadUserDto;
 import com.ddang.ddang.user.presentation.util.NameProcessor;
 
-public record ReadUserResponse(String name, String profileImage, double reliability) {
+public record ReadUserResponse(String name, String profileImage, Float reliability) {
 
     public static ReadUserResponse from(final ReadUserDto readUserDto) {
         return new ReadUserResponse(
